+++ Subastas-app/subastas-api/src/main/java/com/subastas/subastas_api/repository/PujaRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Optional;
 
-public interface PujaRepository extends JpaRepository<Puja, Integer> {
+public interface PujaRepository extends JpaRepository<Puja, Long> {
 
     @Query("""
             SELECT p
